Rename `Logger.js` to `logger.js` (#157)
* Rename `Logger.js` to `logger.js`

* Replaced reference to `logger.js`
