+++ modules/functions.js
@@ -1,4 +1,4 @@
-const logger = require("./Logger.js");
+const logger = require("./logger.js");
 const config = require("../config.js");
 const { settings } = require("./settings.js");
 // Let's start by getting some useful functions that we'll use throughout
